Rename typelevel package to compiletime
