+++ library/src-bootstrapped/scala/Tuple.scala
@@ -1,6 +1,6 @@
 package scala
 import annotation.showAsInfix
-import typelevel._
+import compiletime._
 import scala.StagedTuple
 
 sealed trait Tuple extends Any {
