+++ library/src-bootstrapped/scala/compiletime/Mirror.scala
@@ -1,4 +1,4 @@
-package scala.typelevel
+package scala.compiletime
 
 /** A generic representation of a case in an ADT
   *  @param  reflected  The common class-speficic part of this mirror
