+++ library/src-bootstrapped/scala/compiletime/Shaped.scala
@@ -1,4 +1,4 @@
-package scala.typelevel
+package scala.compiletime
 
 /** Every generic derivation starts with a typeclass instance of this type.
   *  It informs that type `T` has shape `S` and also implements runtime reflection on `T`.
