+++ tests/pos-special/typeclass-scaling.scala
@@ -213,8 +213,8 @@ object typeclasses {
   }
 
   object Eq {
-    import scala.typelevel.erasedValue
-    import typelevel._
+    import scala.compiletime.erasedValue
+    import compiletime._
 
     inline def tryEql[T](x: T, y: T) = implicit match {
       case eq: Eq[T] => eq.eql(x, y)
@@ -268,8 +268,8 @@ object typeclasses {
   }
 
   object Pickler {
-    import scala.typelevel.{erasedValue, constValue}
-    import typelevel._
+    import scala.compiletime.{erasedValue, constValue}
+    import compiletime._
 
     def nextInt(buf: mutable.ListBuffer[Int]): Int = try buf.head finally buf.trimStart(1)
