+++ tests/run/typeclass-derivation1.scala
@@ -1,5 +1,5 @@
 object Deriving {
-  import scala.typelevel._
+  import scala.compiletime._
 
   sealed trait Shape
